separated cloneDeep and cloneDeepInternal
